completed lab7

+++ submit/lab7-sol/multiple-defs/f.c
@@ -1,5 +1,4 @@
-extern double x;
-
+extern int x;
 void f() {
   x = -0.0;
 }
